Move clients into the same file to address circular import issue (they both depend on each other). Missed one case of passing file_database_client to FileCollectionClient constructor.
