+++ tethysext/atcore/models/file_database/__init__.py
@@ -7,7 +7,8 @@
 ********************************************************************************
 """
 from tethysext.atcore.models.file_database.file_collection import FileCollection  # noqa: F401, E402, E501
-from tethysext.atcore.models.file_database.file_collection_client import FileCollectionClient  # noqa: F401, E402, E501
 from tethysext.atcore.models.file_database.file_database import FileDatabase  # noqa: F401, E402, E501
-from tethysext.atcore.models.file_database.file_database_client import FileDatabaseClient  # noqa: F401, E402, E501
-from tethysext.atcore.models.file_database.file_collection_mixin import resource_file_collection_assoc, FileCollectionMixin  # noqa: F401, E402, E501
+from tethysext.atcore.models.file_database.file_database_client import FileDatabaseClient, \
+    FileCollectionClient  # noqa: F401, E402, E501
+from tethysext.atcore.models.file_database.file_collection_mixin import resource_file_collection_assoc, \
+    FileCollectionMixin  # noqa: F401, E402, E501
